Simple racetrack simulator
This is a dog racetrack simulator. There are three bettors and 4 dogs.
The bettors place bets before the race begins. Only one bettor can bet
at a time and can change their but up until the race begins. Once the
race is over if one of the bettors picked the winning dog, the bettor
collects his winnings and the other bettors lose the amount that they
betted